\name{FeedrRewrite-package}
\alias{FeedrRewrite-package}
\alias{FeedrRewrite}
\docType{package}
\title{
  FeedrRewrite
}
\description{
  A small rewrite of the Feedr package into C++ leveraging Rcpp
}
\details{
  Feedr Function | FeedrRewrite Function
		disp	 | displacements
		_	     | encounters
}
\author{
Jimmy Barnes

Maintainer: Jimmy Barnes github/joebart457
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  https://github.com/animalnexus/feedr
  https://www.rcpp.org/
}
\examples{
  \dontrun{
     
  }
}
